requirements.txt
# AWS SDK and cloud services
boto3
botocore

# Web framework and API
fastapi
uvicorn[standard]
pydantic

# Data processing and analysis
pandas
numpy

# Machine learning and model serialization
joblib
scikit-learn

# Environment and configuration
python-dotenv

# Logging (built-in, but good to be explicit about versions)
# logging is part of standard library

# Type hints and utilities (Python 3.8+ compatibility)
typing-extensions
